backend/requirements.txt
Flask==2.3.0
flask-cors==4.0.0
web3==6.0.0
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
werkzeug==2.3.0

